Remove settings
The `trusted_users` and `trusted_groups` settings is effective only when
one of these trusted users creates a Pod object.

Creating a Pod through a higher level resource (like Deployment,
StatefulSet,...) leads to the creation of a Pod using a different user:
an internal Kubernetes service account is used instead, not the very
same user who created the parent object.