Fallback to allowed_profiles to validate securityContext.
When the user defined the seccomp profile using annotations, later
Kubernetes versions will populate the securityContext field
automatically with the equivalent configuration. This can cause
rejections when the users are using the seccomp policy with no
`profile_types` and `localhost_profiles` configuration.

To mitigate this, the policy now fall-back to the `allowed_profiles`
settings when it cannot validate the securityContext configuration with
the `localhost_profiles` and `localhost_profiles` settings.