refactor: improve settings validation
On the `main` branch a set of changes landed that make it possible to
have real settings validation for certificate based signatures.

This commit fixes the build errors and organizes the settins code in a
more structured way.

These are the changes introduced:

  * Have each type of Signature implement the `Display` trait. This is
    used when propagating errors
  * Add a `validate` method to the `Signature` class, then leave each
    enum to implement it. This IMHO makes the validate_settings code
    more maintainable
  * Introduce usage of mocks to stub the crypto host callback functions

About the latter one, the previous settings validation tests were simulating
a failure of certificate verification. The failure was happening, but
not because it was triggered by a mock. Since there was no mocking in
place, the failure was caused by the waPC runtime.
This is now addressed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>