fix: allow resource settings with empty values.
When the policy is used in the Rancher UI the policy settings are sent
with empty values when the user does not want to validate a resource
type (memory/cpu). This is causing issues because the policy does not
allow empty values in the settings. This commit fix that by allowing
only a single resource settings to be empty. Not both. Therefore, we fix
the issue in the Rancher UI and does not broken the policy usage.

The Rancher UI cannot be fixed in the questions.yaml file because it
does not have a feature that allow us to remove the boolean flag when
user does not enabled it. Therefore, the settings always sent empty
values.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>