fix(metadata): revert version
Our CI to release the policies automatically updates the version in the metadata.yml. If this value is changed manually, the CI will not trigger the release PR.

This commit rollback the version in the metadata.yml to allow the CI to start the release process based on the version in the current draft release. 

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>