fix(test): update certificates used to test cert chain.
After a recent change in the Kubewarden host capability used to verify
certificates chain, the verification turn out more restrict. Which
causes the certificates used in the test to validate the certificate
chain signature to fail due another reason. Therefore, the certificate
used by this test have been regenerated.

Before the regeneration of the certificates, the error message from the
certificate validation was about expired certificates. Which are true,
the certificates are expired. Now, the error message is the expected one
about a certificate not signed by the expected certificate chain.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>