chore: update to latest version of SBOMbastic
Update to SBOMbastic 0.6.0. Several breaking changes happened during the
last versions of SBOMbastic, because of that the CRD definitions had to
be updated.

The new `VulnerabilityReport` doesn't have a `.spec` anymore. That means
we cannot use the `k8s-openapi-derive` helpers anymore, since they all
expect to generate an resource that has a `.spec`.

All the fixtures have been updated to reflect the breaking changes done
to the `VulnerabilityReport` CRD.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>