fix: support multi-arch images
Prior to this commit, multi-arch images were not supported by the
policy. That's because the policy looked up the `VulnerabilityReport`
using the digest of the image index, not of the manifest of an image.

With this commit, the policy is not getting the manifest of an image and
then computes the digest accordingly.

When a single architecture image is found, the policy will use the
digest of the only platform that exists.

When a multi-arch image is found, the policy will validate the image
that has the platform equal to the one specified inside of the policy
settings.

The user can also omit to specify the platform inside of the policy's
settings. When that happens, the policy will use the same platform as
the node running the Policy Server instance that hosts the policy.

Warning: currently we cannot determine the platform of the host system.
We need a host capability for that. In the meantime, we hard code the
amd64/linux platform.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>