fix(metadata): remove Nicolas from maintainer list.
Removes Nicolas from the maintainers list because that will be used to
generate the maintainers list for ArtifactHub. As Nicolas is not a
maintainer of this repository, keep his name there is wrong and do not
reflect the actual state of things.

His name is kept as author of the policy in the Cargo.toml file to
give him the credit as the original author of the policy from which we
forked it.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>